requirements.txt
requests
beautifulsoup4
azure-functions